Move some functions to a more sensible place, stop duping ping
Rename get_socket() to be less confusable
