+++ spinnman/connections/udp_packet_connections/bmp_connection.py
@@ -16,10 +16,10 @@
 import struct
 from spinn_utilities.overrides import overrides
 from .udp_connection import UDPConnection
-from .utils import update_sdp_header_for_udp_send
 from spinnman.constants import SCP_SCAMP_PORT
 from spinnman.messages.scp.enums import SCPResult
 from spinnman.connections.abstract_classes import SCPReceiver, SCPSender
+from .utils import update_sdp_header_for_udp_send
 
 _TWO_SHORTS = struct.Struct("<2H")
 _TWO_SKIP = struct.Struct("<2x")
